+++ packages/calcite-components/src/components/alert/alert.e2e.ts
@@ -28,11 +28,11 @@ describe("calcite-alert", () => {
     hidden("<calcite-alert open></calcite-alert>");
   });
 
-  describe("accessible", () => {
+  describe.skip("accessible", () => {
     accessible(html` <calcite-alert open label="test"> ${alertContent} </calcite-alert> `);
   });
 
-  describe("accessible with auto-close", () => {
+  describe.skip("accessible with auto-close", () => {
     accessible(html`
       <calcite-alert open auto-close auto-close-duration="slow" label="test"> ${alertContent} </calcite-alert>
     `);