+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.0-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.9.0-next.0...@esri/calcite-components-react@1.9.0-next.1) (2023-09-18)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.9.0-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.8.1-next.1...@esri/calcite-components-react@1.9.0-next.0) (2023-09-16)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.8.1-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.8.1-next.0...@esri/calcite-components-react@1.8.1-next.1) (2023-09-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.8.1-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.8.0...@esri/calcite-components-react@1.8.1-next.0) (2023-09-08)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [1.8.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0...@esri/calcite-components-react@1.8.0) (2023-09-06)
 
 ### Bug Fixes